Route agent lineage links diagonally
